requirements.txt
numpy==1.23.2
opencv-python-headless==4.8.0.74
Pillow==9.2.0
streamlit==1.24.1
tensorflow==2.13.0
tqdm==4.64.0
